Allow decimal labor hours (Issue #25)
Convert labor.hours from integer to float with 2-decimal-place limit.
Parts keep integer quantities; labor hours now support fractional
values like 1.5, 0.25, 8.75.

+++ backend/models.py
@@ -106,7 +106,7 @@ class Labor(Base):
 
     id = Column(Integer, primary_key=True, index=True)
     description = Column(String, nullable=False)
-    hours = Column(Integer, nullable=False, default=1)  # Must be whole number
+    hours = Column(Float, nullable=False, default=1)
     rate = Column(Float, nullable=False)
     markup_percent = Column(Float, default=0.0)
     category_id = Column(Integer, ForeignKey('categories.id'))
